Rewrite completion responses from Augment so Emacs accepts them
Augment responses to textDocument/completion requests for a completion
to, say, a current token of "requi" are of the form:

    {
        'label': '7a6da89e-7ba0-4212-b183-773328bd5545',
        'insertText': "re 'markdown-view)"
    }

However, per the LSP spec <https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocument_completion>:

    Completion item provides an insertText / label without a text edit: in
    the model the client should filter against what the user has already
    typed using the word boundary rules of the language (e.g. resolving the
    word under the cursor position). The reason for this mode is that it
    makes it extremely easy for a server to implement a basic completion
    list and get it filtered on the client.

and in fact for a response like this, emacs lsp-mode will go into
lsp-completion--guess-prefix and find that the completion does not
match anything and therefore discard it.

Add advice around the request function so that responses are rewritten
to an explicit textEdit like:

    {
        'label': "re 'markdown-view)",
        'textEdit': {
	    'range': {
	        'start': {'line': 15, 'character': 6},
	        'end': {'line': 15, 'character': 6}
	    },
	    'newText': "re 'markdown-view)"
        }
    }

Also update the "label" field to match the proposed completion,
since (again per the LSP spec) that is what the emacs UI uses to
display the proposed completion.