Allow Augment client to coexist with other LSP clients
Correct my misunderstanding of how to integrate the Augment client
with LSP mode: rather than creating a new language "augment," just
have a customizable activation-fn that determines whether to connect
the Augment client for any buffer using lsp-mode. This allows the
Augment client to be used in conjuction with other language-specific
clients.